refactor(fcp): extract client put helpers
Split ClientPut validation, data prep, and status snapshot work into focused helpers and simplify ClientPutPutterFactory to use request/options records.
